+++ packages/model-bank/src/aiModels/taichu.ts
@@ -1,4 +1,4 @@
-import { AIChatModelCard } from '@/types/aiModel';
+import { AIChatModelCard } from '../types/aiModel';
 
 // https://docs.wair.ac.cn/maas/jiage.html