+++ packages/model-runtime/src/utils/openaiCompatibleFactory/createImage.ts
@@ -1,9 +1,9 @@
-import { imageUrlToBase64 } from '@lobechat/utils';
 import createDebug from 'debug';
 import { RuntimeImageGenParamsValue } from 'model-bank';
 import OpenAI from 'openai';
 
 import { CreateImagePayload, CreateImageResponse } from '../../types/image';
+import { imageUrlToBase64 } from '../imageToBase64';
 import { convertImageUrlToFile } from '../openaiHelpers';
 import { parseDataUri } from '../uriParser';