+++ packages/model-runtime/src/utils/openaiHelpers.ts
@@ -1,9 +1,8 @@
-import { imageUrlToBase64 } from '@lobechat/utils';
 import OpenAI, { toFile } from 'openai';
 
-import { disableStreamModels, systemToUserModels } from '@/const/models';
-import { ChatStreamPayload, OpenAIChatMessage } from '@/libs/model-runtime';
-
+import { disableStreamModels, systemToUserModels } from '../const/models';
+import { ChatStreamPayload, OpenAIChatMessage } from '../types';
+import { imageUrlToBase64 } from './imageToBase64';
 import { parseDataUri } from './uriParser';
 
 export const convertMessageContent = async (