+++ packages/types/src/index.ts
@@ -1,5 +1,8 @@
 export * from './agent';
+export * from './aiProvider';
 export * from './artifact';
+export * from './asyncTask';
+export * from './auth';
 export * from './chunk';
 export * from './clientDB';
 export * from './eval';
@@ -8,6 +11,7 @@ export * from './knowledgeBase';
 export * from './llm';
 export * from './message';
 export * from './meta';
+export * from './serverConfig';
 export * from './user';
 export * from './user/settings';
 // FIXME: I think we need a refactor for the "openai" types