+++ src/features/Conversation/Extras/Usage/UsageDetail/ModelCard.tsx
@@ -9,7 +9,7 @@ import { Flexbox } from 'react-layout-kit';
 import { getPrice } from '@/features/Conversation/Extras/Usage/UsageDetail/pricing';
 import { useGlobalStore } from '@/store/global';
 import { systemStatusSelectors } from '@/store/global/selectors';
-import { LobeDefaultAiModelListItem } from '@/types/aiModel';
+import { LobeDefaultAiModelListItem } from '../../../../../../packages/model-bank/src/types/aiModel';
 import { getCachedTextInputUnitRate, getWriteCacheInputUnitRate } from '@/utils/pricing';
 
 export const useStyles = createStyles(({ css, token }) => {